Split simpletests.js into topic-based test files
Same 85 tests and 16 suites, relocated verbatim:
- basic-serialization.js: arrays, objects, primitives, top-level scalars
- numbers.js: NaN, Infinity, number formatting boundaries
- strings.js: unicode handling, escaping, fast-path boundaries
- key-sorting.js: code-unit ordering, sort fallback
- tojson.js: toJSON behavior
- circular-and-depth.js: cycle detection, deep nesting
- non-json-values.js: functions/undefined/symbols, boxed primitives,
  unusual objects
